LayoutTests/imported/w3c/web-platform-tests/xhr/xmlhttprequest-timeout-synconmain-expected.txt

CONSOLE MESSAGE: line 316: XMLHttpRequest.timeout cannot be set for synchronous HTTP(S) requests made from the window context.
CONSOLE MESSAGE: line 290: Synchronous XMLHttpRequests must not have a timeout value set.
Description

This test validates that the XHR2 timeout property behaves as expected in async cases in document (i.e. non-worker) context.


PASS Timeout test: Synchronous XHR must not allow a timeout to be set - setting timeout must throw 
PASS Timeout test: Synchronous XHR must not allow a timeout to be set - calling open() after timeout is set must throw 
